Refactor RouteInterceptor.
