+++ router/src/main/java/com/chenenyu/router/MatcherRegistry.java
@@ -1,4 +1,10 @@
-package com.chenenyu.router.matcher;
+package com.chenenyu.router;
+
+import com.chenenyu.router.matcher.AbsMatcher;
+import com.chenenyu.router.matcher.BrowserMatcher;
+import com.chenenyu.router.matcher.DirectMatcher;
+import com.chenenyu.router.matcher.ImplicitMatcher;
+import com.chenenyu.router.matcher.SchemeMatcher;
 
 import java.util.ArrayList;
 import java.util.Collections;
